+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # ChangeLog
 
 ### 3.4.16
+- CHANGE: #5248 Database Export dialog has now Export Indexes and Export Triggers options, just like the Single Table Export dialog.
 - BUGFIX: #5247 Fixed crash when opening a View's Trigger and enabled trigger-related entries in context menu for a View's Trigger.
 - BUGFIX: #5233 Fixed result column name when using [] quoting operator for column alias name.
 - BUGFIX: #5226 Fixed SQL formatter for INSERT statement when database prefix is used, so the dot is not removed after database name.